requirements.txt
Flask==3.0.2
flask-cors==4.0.0
Pillow==10.2.0
onnxruntime==1.17.1
transformers==4.38.2
huggingface-hub==0.21.4
supabase==2.3.4
gunicorn==21.2.0
